Extract GreenRedBranch.kt
